From Vadim Peretokin <vperetokin@gmail.com>:
The Valencia plugin was failing to load twice because of re-registering stuff.
The issue was that several classes did not derive from Glib.Object and as such
were "fundamental" - which does not work in a plugin.

So this is a patch that fixes this and makes the plugin activatable however
many times you want. There still are some warnings, but Jürg Billeter from
Vala said he'll look into why are those happening (it's a vala thing).