Manually wipe Valencia's loaded symbols
Because Valencia can't detect when a file has been changed outside of
gedit or renamed, it's possible for its symbol table to get out of
sync with the actual file contents.  A better solution would be to
detect these changes and reload, but until that occurs, this menu item
allows for all symbols to be wiped from memory.  The next time the
user does a symbol lookup, Valencia will reload all symbols (much as
if the user had quit and restart gedit).